+++ devdesk/src/App.js
@@ -2,14 +2,14 @@ import React from "react";
 import "./App.css";
 import { Route, Switch } from "react-router-dom";
 import Login from "./Components/Login";
-import Queue from "./Components/Queue";
+import IssueList from "./Components/IssueList";
 import Registar from "./Components/Registar";
 
 function App() {
   return (
     <div className="App">
       <Switch>
-        <Route path="/queue" component={Queue} />
+        <Route path="/issues" component={IssueList} />
         <Route path="/registar" component={Registar} />
         <Route exact path="/" component={Login} />
       </Switch>
